# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "Accion": {
        "enum": [
          "ejecutar",
          "editar-insumos",
          "congelar",
          "recalcular"
        ],
        "title": "Accion",
        "type": "string"
      },
...
      "CorridaCongelada": {
        "properties": {
          "congelada_en": {
            "format": "date-time",
            "title": "Congelada En",
            "type": "string"
          },
          "congelada_por": {
            "title": "Congelada Por",
            "type": "string"
          },
          "huella_contenido": {
            "title": "Huella Contenido",
            "type": "string"
          },
          "huella_sello": {
            "title": "Huella Sello",
            "type": "string"
          },
          "id_corrida": {
            "title": "Id Corrida",
            "type": "string"
          },
          "ruta_imagen": {
            "description": "Ubicación de la imagen sellada",
            "title": "Ruta Imagen",
            "type": "string"
          }
        },
        "required": [
          "id_corrida",
          "huella_contenido",
          "huella_sello",
          "congelada_por",
          "congelada_en",
          "ruta_imagen"
        ],
        "title": "CorridaCongelada",
        "type": "object"
      },
      "DesviacionLinea": {
        "properties": {
          "anio": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Anio"
          },
          "dentro_de_tolerancia": {
            "title": "Dentro De Tolerancia",
            "type": "boolean"
          },
          "diferencia_relativa": {
            "title": "Diferencia Relativa",
            "type": "number"
          },
          "linea": {
            "title": "Linea",
            "type": "string"
          },
          "valor_plataforma": {
            "title": "Valor Plataforma",
            "type": "number"
          },
          "valor_referencia": {
            "title": "Valor Referencia",
            "type": "number"
          }
        },
        "required": [
          "linea",
          "valor_plataforma",
          "valor_referencia",
          "diferencia_relativa",
          "dentro_de_tolerancia"
        ],
        "title": "DesviacionLinea",
        "type": "object"
      },
      "DetalleCaso": {
        "properties": {
          "acciones_disponibles": {
            "items": {
              "$ref": "#/components/schemas/Accion"
            },
            "title": "Acciones Disponibles",
            "type": "array"
          },
          "actualizado_en": {
            "format": "date-time",
            "title": "Actualizado En",
            "type": "string"
          },
          "actualizado_por": {
            "title": "Actualizado Por",
            "type": "string"
          },
          "descripcion": {
            "title": "Descripcion",
            "type": "string"
          },
          "estado": {
            "$ref": "#/components/schemas/Estado"
          },
          "id_caso": {
            "title": "Id Caso",
            "type": "string"
          },
          "nombre": {
            "title": "Nombre",
            "type": "string"
          },
          "terna": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Terna"
              },
              {
                "type": "null"
              }
            ],
            "description": "Presente solo si el caso tiene una corrida"
          },
          "tipo": {
            "enum": [
              "sin-proyecto",
              "monometalico",
              "polimetalico"
            ],
            "title": "Tipo",
            "type": "string"
          }
        },
        "required": [
          "id_caso",
          "nombre",
          "tipo",
          "estado",
          "actualizado_en",
          "actualizado_por",
          "descripcion",
          "acciones_disponibles"
        ],
        "title": "DetalleCaso",
        "type": "object"
      },
...
      "Estado": {
        "enum": [
          "borrador",
          "calculada",
          "congelada"
        ],
        "title": "Estado",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Indicadores": {
        "properties": {
          "capital_intensity": {
            "title": "Capital Intensity",
            "type": "number"
          },
          "npv_musd": {
            "description": "Valor actual neto, millones de dólares",
            "title": "Npv Musd",
            "type": "number"
          },
          "payback_anios": {
            "title": "Payback Anios",
            "type": "number"
          },
          "tir": {
            "description": "Tasa interna de retorno, en tanto por uno",
            "title": "Tir",
            "type": "number"
          }
        },
        "required": [
          "npv_musd",
          "tir",
          "payback_anios",
          "capital_intensity"
        ],
        "title": "Indicadores",
        "type": "object"
      },
...
      "Problema": {
        "description": "Cuerpo de error. Explica qué pasó y cómo resolverlo.",
        "properties": {
          "detalle": {
            "title": "Detalle",
            "type": "string"
          },
          "restriccion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Restricción del cuadro de control que bloquea, si aplica",
            "title": "Restriccion"
          }
        },
        "required": [
          "detalle"
        ],
        "title": "Problema",
        "type": "object"
      },
      "ResultadoEvaluacion": {
        "properties": {
          "duracion_ms": {
            "title": "Duracion Ms",
            "type": "integer"
          },
          "ejecutada_en": {
            "format": "date-time",
            "title": "Ejecutada En",
            "type": "string"
          },
          "ejecutada_por": {
            "title": "Ejecutada Por",
            "type": "string"
          },
          "estado": {
            "$ref": "#/components/schemas/Estado"
          },
          "id_caso": {
            "title": "Id Caso",
            "type": "string"
          },
          "id_corrida": {
            "title": "Id Corrida",
            "type": "string"
          },
          "indicadores": {
            "$ref": "#/components/schemas/Indicadores"
          },
          "terna": {
            "$ref": "#/components/schemas/Terna"
          }
        },
        "required": [
          "id_corrida",
          "id_caso",
          "estado",
          "terna",
          "indicadores",
          "ejecutada_en",
          "ejecutada_por",
          "duracion_ms"
        ],
        "title": "ResultadoEvaluacion",
        "type": "object"
      },
...
      "ResultadoFidelidad": {
        "description": "Contraste contra el modelo de referencia en los cuatro niveles.",
        "properties": {
          "dentro_de_tolerancia": {
            "title": "Dentro De Tolerancia",
            "type": "boolean"
          },
          "desviaciones": {
            "items": {
              "$ref": "#/components/schemas/DesviacionLinea"
            },
            "title": "Desviaciones",
            "type": "array"
          },
          "id_caso": {
            "title": "Id Caso",
            "type": "string"
          },
          "nivel_n0": {
            "description": "Paridad de inputs, sin tolerancia",
            "title": "Nivel N0",
            "type": "boolean"
          },
          "nivel_n1": {
            "description": "Bloques intermedios, año a año",
            "title": "Nivel N1",
            "type": "boolean"
          },
          "nivel_n2": {
            "description": "Flujos, año a año",
            "title": "Nivel N2",
            "type": "boolean"
          },
          "nivel_n3": {
            "description": "Indicadores finales",
            "title": "Nivel N3",
            "type": "boolean"
          },
          "version_motor": {
            "title": "Version Motor",
            "type": "string"
          }
        },
        "required": [
          "id_caso",
          "version_motor",
          "dentro_de_tolerancia",
          "nivel_n0",
          "nivel_n1",
          "nivel_n2",
          "nivel_n3"
        ],
        "title": "ResultadoFidelidad",
        "type": "object"
      },
...
      "SolicitudCongelamiento": {
        "properties": {
          "motivo": {
            "description": "Qué decisión sustenta esta evaluación. Es lo que explica, años después, por qué se congeló",
            "maxLength": 500,
            "minLength": 10,
            "title": "Motivo",
            "type": "string"
          },
          "rutas_respaldo": {
            "description": "Documentos y correos de aprobación asociados",
            "items": {
              "type": "string"
            },
            "title": "Rutas Respaldo",
            "type": "array"
          }
        },
        "required": [
          "motivo"
        ],
        "title": "SolicitudCongelamiento",
        "type": "object"
      },
...
      "Terna": {
        "description": "Las tres versiones que definen una corrida.",
        "properties": {
          "datos_maestros": {
            "description": "Comité de Precios utilizado",
            "examples": [
              "CP-2026-03"
            ],
            "title": "Datos Maestros",
            "type": "string"
          },
          "huella_inputs": {
            "maxLength": 64,
            "minLength": 64,
            "title": "Huella Inputs",
            "type": "string"
          },
          "motor": {
            "description": "Versión del motor de cálculo",
            "examples": [
              "1.2.0"
            ],
            "title": "Motor",
            "type": "string"
          },
          "revision_inputs": {
            "minimum": 1.0,
            "title": "Revision Inputs",
            "type": "integer"
          },
          "vigencia_datos_maestros": {
            "format": "date",
            "title": "Vigencia Datos Maestros",
            "type": "string"
          }
        },
        "required": [
          "motor",
          "datos_maestros",
          "vigencia_datos_maestros",
          "revision_inputs",
          "huella_inputs"
        ],
        "title": "Terna",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "contact": {
      "name": "INVA · Servicio INVA-01-2026-182"
    },
    "description": "\nServicios de la Plataforma de Evaluación Económica · Modelo MINSUR.\n\nReproduce la lógica del modelo económico corporativo para que el área de\nProyectos evalúe escenarios de forma autónoma, con trazabilidad completa.\n\n**Autenticación.** Token de Microsoft Entra ID del tenant corporativo. El\nperfil se deriva de la reclamación `groups`; la plataforma no consulta la\nmembresía a Microsoft Graph.\n\n**Archivos.** Las plantillas se cargan y los reportes se descargan\ndirectamente contra el almacenamiento, mediante firmas de acceso compartido\nde corta vigencia. No atraviesan esta API.\n\n**Respuestas 501.** Indican una operación cuya implementación depende de un\ninsumo del cliente aún no recibido. El cuerpo incluye la restricción del\ncuadro de control que la bloquea.\n",
    "title": "Plataforma de Evaluación Económica · MINSUR",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/casos/{id_caso}": {
      "get": {
        "description": "Incluye las acciones disponibles para el perfil del solicitante.",
        "operationId": "detalle_api_casos__id_caso__get",
        "parameters": [
          {
            "in": "path",
            "name": "id_caso",
            "required": true,
            "schema": {
              "examples": [
                "CASO-SR-2026-014"
              ],
              "title": "Id Caso",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleCaso"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Detalle de un caso",
        "tags": [
          "Casos"
        ]
      }
    },
...
    "/api/casos/{id_caso}/congelar": {
      "post": {
        "description": "Transición irreversible.\n\nGenera una imagen sellada autocontenida —inputs, parámetros efectivamente\nutilizados, versión del motor, resultados y documentación de respaldo—,\nla resume con SHA-256 y la deposita en el contenedor con política de\ninmutabilidad.\n\nDesde ese momento la evaluación no puede alterarse. Para ver el efecto de\nun cambio posterior del modelo, se recalcula: eso genera una corrida\nnueva y deja esta intacta como sustento.\n\nCongelar corresponde a quien sustenta la decisión, no a quien opera la\nplataforma. Por eso el Ingeniero de Proyecto no aparece entre los\nperfiles autorizados.",
        "operationId": "congelar_api_casos__id_caso__congelar_post",
        "parameters": [
          {
            "description": "Identificador del caso",
            "in": "path",
            "name": "id_caso",
            "required": true,
            "schema": {
              "description": "Identificador del caso",
              "examples": [
                "CASO-SR-2026-014"
              ],
              "title": "Id Caso",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitudCongelamiento"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorridaCongelada"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Congelar la evaluación como sustento de decisión",
        "tags": [
          "Evaluación"
        ]
      }
    },
    "/api/casos/{id_caso}/evaluar": {
      "post": {
        "description": "Calcula la evaluación y deja la corrida en estado *calculada*.\n\nRegistra la terna de versiones con la que se calculó. Una corrida no\nguarda una referencia a la versión vigente del motor: guarda la versión\nexacta, para que publicar una versión nueva no altere corridas previas.",
        "operationId": "evaluar_api_casos__id_caso__evaluar_post",
        "parameters": [
          {
            "description": "Identificador del caso",
            "in": "path",
            "name": "id_caso",
            "required": true,
            "schema": {
              "description": "Identificador del caso",
              "examples": [
                "CASO-SR-2026-014"
              ],
              "title": "Id Caso",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoEvaluacion"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Ejecutar la evaluación económica",
        "tags": [
          "Evaluación"
        ]
      }
    },
...
    "/api/casos/{id_caso}/recalcular": {
      "post": {
        "description": "Crea una corrida nueva a partir de una congelada. **No la modifica.**\n\nEs la forma de observar el efecto de un cambio del modelo sin que la\nevidencia que sustentó una decisión deje de existir. La corrida nueva\napunta a la anterior como origen.",
        "operationId": "recalcular_api_casos__id_caso__recalcular_post",
        "parameters": [
          {
            "description": "Identificador del caso",
            "in": "path",
            "name": "id_caso",
            "required": true,
            "schema": {
              "description": "Identificador del caso",
              "examples": [
                "CASO-SR-2026-014"
              ],
              "title": "Id Caso",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoEvaluacion"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recalcular una corrida congelada con el motor vigente",
        "tags": [
          "Evaluación"
        ]
      }
    },
    "/api/casos/{id_caso}/verificar-fidelidad": {
      "get": {
        "description": "Contraste en los cuatro niveles.\n\n| Nivel | Objeto | Criterio |\n|---|---|---|\n| N0 | Paridad de inputs | Coincidencia exacta |\n| N1 | Bloques intermedios, año a año | 0,1 % o US$ 10 000 por línea |\n| N2 | Flujos, año a año | 0,1 % o US$ 10 000 por año |\n| N3 | Indicadores finales | NPV 0,1 % · TIR 5 pb · Payback 0,1 año |\n\nVerificar solo el indicador final es metodológicamente insuficiente: dos\nerrores compensatorios en bloques intermedios producen un NPV correcto\nsobre un modelo inválido.\n\nLos umbrales son la propuesta de INVA; el valor contractual lo fija\nFinanzas (`R-31`). Este endpoint es el que consulta la segunda de las\ncinco comprobaciones obligatorias de la ventana de pase.",
        "operationId": "verificar_fidelidad_api_casos__id_caso__verificar_fidelidad_get",
        "parameters": [
          {
            "description": "Identificador del caso",
            "in": "path",
            "name": "id_caso",
            "required": true,
            "schema": {
              "description": "Identificador del caso",
              "examples": [
                "CASO-SR-2026-014"
              ],
              "title": "Id Caso",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "default": "",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultadoFidelidad"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            },
            "description": "Not Implemented"
          }
        },
        "summary": "Contrastar contra el modelo de referencia",
        "tags": [
          "Evaluación"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "A través de API Management",
      "url": "/"
    }
  ]
}